feat: refactor some css
