--- src/Components/HamburgurMenu/HamburgurMenu.css
@@ -1,25 +1,0 @@
-:root {
-  --light-color: #eff6ff;
-}
-.hamburgur-menu {
-  padding: 1rem;
-  position: fixed;
-  z-index: 3;
-  top: 3rem;
-  left: 0;
-  list-style: none;
-  background-color: var(--light-color);
-  width: 100%;
-  display: flex;
-  flex-direction: column;
-  transform: 0.3s ease-in-out;
-}
-.hamburgur-menu-item {
-  color: black;
-  margin-bottom: 1.5rem;
-  font-weight: bold;
-}
-.hamburgur-menu-item:hover {
-  cursor: pointer;
-  color: var(--medium-blue);
-}
