+++ src/Components/Index.js
@@ -4,6 +4,5 @@ export { ProductCard } from "./ProductCard/ProductCard";
 export { Categories } from "./Categories/Categories";
 export { ProductCardForCart } from "./CartCards/ProductCardForCart";
 export { BillCardForCart } from "./CartCards/BillCardForCart";
-export { HamburgurMenuList } from "./HamburgurMenu/HamburgurMenu";
 export { Footer } from "./Footer/Footer";
 export { NavbarMblView } from "./NavbarMblView/NavbarMblView";
